spark/requirements.txt
# Producer
kafka-python
pyzmq

# Consumer
pyspark=4.0.1
pandas

# OLAP
duckdb
